INBOX 073: retract the hgcc-only theory; the box was building a stale generated table
The box tree matched HEAD on every source file and on the actlize submodule, but
benchmarks/lowbit_dense_configs.inc was an uncommitted stale regeneration: 150 rows
against HEAD 227, and ZERO overlap in (TileM,TileN,TileK,WarpM) despite identical
header comments and field count. So every compile error forwarded so far came from
configs that are not in the current tactic space, and the nvcc-vs-hgcc comparison in
072 was never run on the same input.

Also carries a portability fix in the dense bench: the brace-init of the shape pair
passed to preprocess_weights_for_mixed_gemm needs explicit size_t casts, since the
ints narrow.

+++ benchmarks/test_lowbit_dense_bench.cu
@@ -617,7 +617,8 @@ void initialize(Options const& options) {
     int64_t batch_offset = b * row * col;
     // preprocess_weights_for_mixed_gemm<is_rowmajor, -1>((int8_t*)(&block_B_buff.host_data()[batch_offset]),
     preprocess_weights_for_mixed_gemm<is_rowmajor, 256>((int8_t*)(&block_B_buff.host_data()[batch_offset]),
-        (int8_t*)(&tensor_B.host_data()[batch_offset]), {col, row}, quant_type);
+        (int8_t*)(&tensor_B.host_data()[batch_offset]),
+        {static_cast<size_t>(col), static_cast<size_t>(row)}, quant_type);
   }
   block_B_buff.sync_device();
 }
